requirements.txt
pandas
apache-airflow
newsapi-python
logging